tot en met 4d week 6

+++ Project Kantine/KantineSimulatieBasis/src/main/java/Artikel.java
@@ -29,7 +29,7 @@ public class Artikel {
         return naam;
     }
 
-    public int getPrijs(){
+    public double getPrijs(){
         return prijs;
     }
